# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UK On Wheels operations portal API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardMetric {
  label: string;
  value: string;
  change: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export type ActivityTone = typeof ActivityTone[keyof typeof ActivityTone];


export const ActivityTone = {
  success: 'success',
  warning: 'warning',
  info: 'info',
} as const;

export interface Activity {
  id: number;
  title: string;
  detail: string;
  occurredAt: string;
  tone: ActivityTone;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  Pending: 'Pending',
  Assigned: 'Assigned',
  Collected: 'Collected',
  In_Transit: 'In Transit',
  Delivered: 'Delivered',
  Cancelled: 'Cancelled',
} as const;

export interface Booking {
  id: number;
  reference: string;
  customer: string;
  vehicle: string;
  registration?: string;
  pickup: string;
  destination: string;
  collectionDate: string;
  status: BookingStatus;
  amount: number;
  driver: string;
  createdAt: string;
}

export interface Dashboard {
  metrics: DashboardMetric[];
  statusCounts: StatusCount[];
  recentBookings: Booking[];
  activity: Activity[];
}

export interface BookingInput {
  /** @minLength 2 */
  customer: string;
  /** @minLength 2 */
  vehicle: string;
  registration?: string;
  /** @minLength 2 */
  pickup: string;
  /** @minLength 2 */
  destination: string;
  collectionDate: string;
  /** @minimum 0 */
  amount: number;
  driver?: string;
}

export type BookingUpdateStatus = typeof BookingUpdateStatus[keyof typeof BookingUpdateStatus];


export const BookingUpdateStatus = {
  Pending: 'Pending',
  Assigned: 'Assigned',
  Collected: 'Collected',
  In_Transit: 'In Transit',
  Delivered: 'Delivered',
  Cancelled: 'Cancelled',
} as const;

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  driver?: string;
  /** @minimum 0 */
  amount?: number;
}

export type ListBookingsParams = {
status?: string;
search?: string;
};