# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UK On Wheels operations portal API
 * OpenAPI spec version: 1.0.0
 */
import type { ActivityTone } from './activityTone';

export interface Activity {
  id: number;
  title: string;
  detail: string;
  occurredAt: Date;
  tone: ActivityTone;
}