# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UK On Wheels operations portal API
 * OpenAPI spec version: 1.0.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  reference: string;
  customer: string;
  vehicle: string;
  registration?: string;
  pickup: string;
  destination: string;
  collectionDate: Date;
  status: BookingStatus;
  amount: number;
  driver: string;
  createdAt: Date;
}