# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UK On Wheels operations portal API
 * OpenAPI spec version: 1.0.0
 */

export interface BookingInput {
  /** @minLength 2 */
  customer: string;
  /** @minLength 2 */
  vehicle: string;
  registration?: string;
  /** @minLength 2 */
  pickup: string;
  /** @minLength 2 */
  destination: string;
  collectionDate: Date;
  /** @minimum 0 */
  amount: number;
  driver?: string;
}