# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UK On Wheels operations portal API
 * OpenAPI spec version: 1.0.0
 */
import type { BookingUpdateStatus } from './bookingUpdateStatus';

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  driver?: string;
  /** @minimum 0 */
  amount?: number;
}