# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UK On Wheels operations portal API
 * OpenAPI spec version: 1.0.0
 */
import type { Activity } from './activity';
import type { Booking } from './booking';
import type { DashboardMetric } from './dashboardMetric';
import type { StatusCount } from './statusCount';

export interface Dashboard {
  metrics: DashboardMetric[];
  statusCounts: StatusCount[];
  recentBookings: Booking[];
  activity: Activity[];
}