# lib/api-zod/src/generated/types/dashboardMetric.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UK On Wheels operations portal API
 * OpenAPI spec version: 1.0.0
 */

export interface DashboardMetric {
  label: string;
  value: string;
  change: string;
}