# lib/api-zod/src/generated/types/statusCount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UK On Wheels operations portal API
 * OpenAPI spec version: 1.0.0
 */

export interface StatusCount {
  status: string;
  count: number;
}